+++ CHANGELOG.md
@@ -4,6 +4,17 @@ Motion adheres to [Semantic Versioning](http://semver.org/).
 
 Undocumented APIs should be considered internal and may change without warning.
 
+## [13.2.0] 2026-unreleased
+
+### Added
+
+-   `animate` for Three.js supporting objects, materials, shader uniforms, TSL uniform nodes and node material slots like `colorNode`.
+-   `animate` for vgpu supporting shader uniforms.
+
+### Changed
+
+-   Reduced filesize and improved performance of `spring`.
+
 ## [13.1.1] 2026-08-18
 
 ### Fixed